@@ -5,11 +5,12 @@ author: Alibaba Xianyu
 homepage: https://github.com/alibaba/flutter_boost
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   flutter_lints: ^1.0.4
