@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_boost/flutter_boost.dart';
 
 class FlutterRebuildDemo extends StatefulWidget {
-  const FlutterRebuildDemo({Key key}) : super(key: key);
+  const FlutterRebuildDemo({Key? key}) : super(key: key);
 
   @override
   _FlutterRebuildDemoState createState() => _FlutterRebuildDemoState();
@@ -51,7 +51,7 @@ class _FlutterRebuildDemoState extends State<FlutterRebuildDemo> {
 }
 
 class FlutterRebuildPageA extends StatefulWidget {
-  const FlutterRebuildPageA({Key key}) : super(key: key);
+  const FlutterRebuildPageA({Key? key}) : super(key: key);
 
   @override
   _FlutterRebuildPageAState createState() => _FlutterRebuildPageAState();
@@ -100,7 +100,7 @@ class _FlutterRebuildPageAState extends State<FlutterRebuildPageA> {
 }
 
 class FlutterRebuildPageB extends StatefulWidget {
-  const FlutterRebuildPageB({Key key}) : super(key: key);
+  const FlutterRebuildPageB({Key? key}) : super(key: key);
 
   @override
   _FlutterRebuildPageBState createState() => _FlutterRebuildPageBState();
