@@ -9,7 +9,7 @@ import 'package:image_picker/image_picker.dart';
 import 'package:video_player/video_player.dart';
 
 class ImagePickerPage extends StatefulWidget {
-  ImagePickerPage({Key key, this.title}) : super(key: key);
+  ImagePickerPage({required this.title});
 
   final String title;
 
@@ -18,18 +18,16 @@ class ImagePickerPage extends StatefulWidget {
 }
 
 class _ImagePickerPageState extends State<ImagePickerPage> {
-  List<XFile> _imageFileList;
 
-  set _imageFile(XFile value) {
-    _imageFileList = value == null ? null : [value];
-  }
+  XFile? _imageFile;
+
 
   dynamic _pickImageError;
   bool isVideo = false;
 
-  VideoPlayerController _controller;
-  VideoPlayerController _toBeDisposed;
-  String _retrieveDataError;
+  VideoPlayerController? _controller;
+  VideoPlayerController? _toBeDisposed;
+  String _retrieveDataError = "";
 
   final ImagePicker _picker = ImagePicker();
   final TextEditingController maxWidthController = TextEditingController();
@@ -61,35 +59,17 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
   }
 
   void _onImageButtonPressed(ImageSource source,
-      {BuildContext context, bool isMultiImage = false}) async {
+      {bool isMultiImage = false}) async {
     if (_controller != null) {
-      await _controller.setVolume(0.0);
+      await _controller!.setVolume(0.0);
     }
     if (isVideo) {
-      final XFile file = await _picker.pickVideo(
+      final XFile? file = await _picker.pickVideo(
           source: source, maxDuration: const Duration(seconds: 10));
-      await _playVideo(file);
-    } else if (isMultiImage) {
-      await _displayPickImageDialog(context,
-          (double maxWidth, double maxHeight, int quality) async {
-        try {
-          final pickedFileList = await _picker.pickMultiImage(
-            maxWidth: maxWidth,
-            maxHeight: maxHeight,
-            imageQuality: quality,
-          );
-          setState(() {
-            _imageFileList = pickedFileList;
-          });
-        } catch (e) {
-          setState(() {
-            _pickImageError = e;
-          });
-        }
-      });
+      await _playVideo(file!);
     } else {
       await _displayPickImageDialog(context,
-          (double maxWidth, double maxHeight, int quality) async {
+          (double? maxWidth, double? maxHeight, int? quality) async {
         try {
           final pickedFile = await _picker.pickImage(
             source: source,
@@ -98,7 +78,7 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
             imageQuality: quality,
           );
           setState(() {
-            _imageFile = pickedFile;
+            _imageFile = pickedFile!;
           });
         } catch (e) {
           setState(() {
@@ -112,8 +92,8 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
   @override
   void deactivate() {
     if (_controller != null) {
-      _controller.setVolume(0.0);
-      _controller.pause();
+      _controller!.setVolume(0.0);
+      _controller!.pause();
     }
     super.deactivate();
   }
@@ -129,17 +109,13 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
 
   Future<void> _disposeVideoController() async {
     if (_toBeDisposed != null) {
-      await _toBeDisposed.dispose();
+      await _toBeDisposed!.dispose();
     }
     _toBeDisposed = _controller;
     _controller = null;
   }
 
   Widget _previewVideo() {
-    final Text retrieveError = _getRetrieveErrorWidget();
-    if (retrieveError != null) {
-      return retrieveError;
-    }
     if (_controller == null) {
       return const Text(
         'You have not yet picked a video',
@@ -148,32 +124,13 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
     }
     return Padding(
       padding: const EdgeInsets.all(10.0),
-      child: AspectRatioVideo(_controller),
+      child: AspectRatioVideo(_controller!),
     );
   }
 
   Widget _previewImages() {
-    final Text retrieveError = _getRetrieveErrorWidget();
-    if (retrieveError != null) {
-      return retrieveError;
-    }
-    if (_imageFileList != null) {
-      return Semantics(
-          child: ListView.builder(
-            key: UniqueKey(),
-            itemBuilder: (context, index) {
-              // Why network for web?
-              // See https://pub.dev/packages/image_picker#getting-ready-for-the-web-platform
-              return Semantics(
-                label: 'image_picker_example_picked_image',
-                child: kIsWeb
-                    ? Image.network(_imageFileList[index].path)
-                    : Image.file(File(_imageFileList[index].path)),
-              );
-            },
-            itemCount: _imageFileList.length,
-          ),
-          label: 'image_picker_example_picked_images');
+    if (_imageFile != null) {
+      return Image.file(File(_imageFile!.path));
     } else if (_pickImageError != null) {
       return Text(
         'Pick image error: $_pickImageError',
@@ -203,16 +160,15 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
     if (response.file != null) {
       if (response.type == RetrieveType.video) {
         isVideo = true;
-        await _playVideo(response.file);
+        await _playVideo(response.file!);
       } else {
         isVideo = false;
         setState(() {
-          _imageFile = response.file;
-          _imageFileList = response.files;
+          _imageFile = response.file!;
         });
       }
     } else {
-      _retrieveDataError = response.exception.code;
+      // _retrieveDataError = response.exception.code;
     }
   }
 
@@ -261,7 +217,7 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
             child: FloatingActionButton(
               onPressed: () {
                 isVideo = false;
-                _onImageButtonPressed(ImageSource.gallery, context: context);
+                _onImageButtonPressed(ImageSource.gallery);
               },
               heroTag: 'image0',
               tooltip: 'Pick Image from gallery',
@@ -275,7 +231,6 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
                 isVideo = false;
                 _onImageButtonPressed(
                   ImageSource.gallery,
-                  context: context,
                   isMultiImage: true,
                 );
               },
@@ -289,7 +244,7 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
             child: FloatingActionButton(
               onPressed: () {
                 isVideo = false;
-                _onImageButtonPressed(ImageSource.camera, context: context);
+                _onImageButtonPressed(ImageSource.camera);
               },
               heroTag: 'image2',
               tooltip: 'Take a Photo',
@@ -327,14 +282,6 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
     );
   }
 
-  Text _getRetrieveErrorWidget() {
-    if (_retrieveDataError != null) {
-      final Text result = Text(_retrieveDataError);
-      _retrieveDataError = null;
-      return result;
-    }
-    return null;
-  }
 
   Future<void> _displayPickImageDialog(
       BuildContext context, OnPickImageCallback onPick) async {
@@ -376,13 +323,13 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
               TextButton(
                   child: const Text('PICK'),
                   onPressed: () {
-                    double width = maxWidthController.text.isNotEmpty
+                    double? width = maxWidthController.text.isNotEmpty
                         ? double.parse(maxWidthController.text)
-                        : null;
-                    double height = maxHeightController.text.isNotEmpty
+                        : 400;
+                    double? height = maxHeightController.text.isNotEmpty
                         ? double.parse(maxHeightController.text)
                         : null;
-                    int quality = qualityController.text.isNotEmpty
+                    int? quality = qualityController.text.isNotEmpty
                         ? int.parse(qualityController.text)
                         : null;
                     onPick(width, height, quality);
@@ -395,7 +342,7 @@ class _ImagePickerPageState extends State<ImagePickerPage> {
 }
 
 typedef void OnPickImageCallback(
-    double maxWidth, double maxHeight, int quality);
+    double? maxWidth, double? maxHeight, int? quality);
 
 class AspectRatioVideo extends StatefulWidget {
   AspectRatioVideo(this.controller);
@@ -437,7 +384,7 @@ class AspectRatioVideoState extends State<AspectRatioVideo> {
     if (initialized) {
       return Center(
         child: AspectRatio(
-          aspectRatio: controller.value?.aspectRatio,
+          aspectRatio: controller.value.aspectRatio,
           child: VideoPlayer(controller),
         ),
       );
