@@ -7,11 +7,11 @@ typedef TextViewCreatedCallback = void Function(TextViewController controller);
 
 class TextView extends StatefulWidget {
   const TextView({
-    Key key,
+    Key? key,
     this.onTextViewCreated,
   }) : super(key: key);
 
-  final TextViewCreatedCallback onTextViewCreated;
+  final TextViewCreatedCallback? onTextViewCreated;
 
   @override
   State<StatefulWidget> createState() => _TextViewState();
@@ -34,7 +34,7 @@ class _TextViewState extends State<TextView> {
     if (widget.onTextViewCreated == null) {
       return;
     }
-    widget.onTextViewCreated(TextViewController._(id));
+    widget.onTextViewCreated!(TextViewController._(id));
   }
 }
 
