@@ -2,9 +2,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class WillPopRoute extends StatefulWidget {
-  WillPopRoute({Key key, this.title}) : super(key: key);
+  WillPopRoute({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   State<StatefulWidget> createState() => new _WillPopRouteState();
