@@ -43,7 +43,7 @@ class AppLifecycleObserver with GlobalPageVisibilityObserver {
 
 ///单个生命周期示例
 class LifecycleTestPage extends StatefulWidget {
-  const LifecycleTestPage({Key key}) : super(key: key);
+  const LifecycleTestPage({Key? key}) : super(key: key);
 
   @override
   _LifecycleTestPageState createState() => _LifecycleTestPageState();
@@ -87,7 +87,7 @@ class _LifecycleTestPageState extends State<LifecycleTestPage>
     super.didChangeDependencies();
 
     ///注册监听器
-    PageVisibilityBinding.instance.addObserver(this, ModalRoute.of(context));
+    PageVisibilityBinding.instance.addObserver(this, ModalRoute.of(context)!);
   }
 
   @override
