@@ -3,7 +3,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter_boost/flutter_boost.dart';
 
 class ReplacementPage extends StatefulWidget {
-  const ReplacementPage({Key key}) : super(key: key);
+  const ReplacementPage({Key? key}) : super(key: key);
 
   @override
   _ReplacementPageState createState() => _ReplacementPageState();
