@@ -8,7 +8,7 @@ class BoostCacheWidget extends StatefulWidget {
   final WidgetBuilder builder;
 
   const BoostCacheWidget(
-      {@required this.uniqueId, @required this.builder, Key key})
+      {required this.uniqueId, required this.builder, Key? key})
       : assert(builder != null),
         assert(uniqueId != null),
         super(key: key);
@@ -18,8 +18,8 @@ class BoostCacheWidget extends StatefulWidget {
 }
 
 class _BoostCacheWidgetState extends State<BoostCacheWidget> {
-  Widget _cacheWidget;
-  String _oldUniqueId;
+  late Widget _cacheWidget;
+  String? _oldUniqueId;
 
   @override
   Widget build(BuildContext context) {
