@@ -4,7 +4,7 @@ import 'container_overlay.dart';
 import 'flutter_boost_app.dart';
 import 'messages.dart';
 
-typedef EventListener = Future<dynamic> Function(String key, Map arguments);
+typedef EventListener = Future<dynamic>? Function(String key, Map arguments);
 
 /// The [BoostChannel] is a tool to get [FlutterBoostAppState]
 /// to operate the Custom events
@@ -16,7 +16,7 @@ class BoostChannel {
   ///The singleton for [BoostChannel]
   static final BoostChannel _instance = BoostChannel._();
 
-  FlutterBoostAppState _appState;
+  FlutterBoostAppState? _appState;
 
   static BoostChannel get instance {
     _instance._appState ??= overlayKey.currentContext
@@ -29,7 +29,7 @@ class BoostChannel {
   ///
   /// The [VoldCallBack] is to remove this listener
   VoidCallback addEventListener(String key, EventListener listener) {
-    return _appState.addEventListener(key, listener);
+    return _appState!.addEventListener(key, listener);
   }
 
   ///Send a custom event to native with [key] and [args]
@@ -37,16 +37,15 @@ class BoostChannel {
   void sendEventToNative(String key, Map args) {
     assert(key != null);
 
-    args ??= {};
 
     var params = CommonParams()
       ..key = key
       ..arguments = args;
-    _appState.nativeRouterApi.sendEventToNative(params);
+    _appState!.nativeRouterApi!.sendEventToNative(params);
   }
 
   /// enable iOS native pop gesture for container matching [containerId]
-  void enablePopGesture({@required String containerId}) {
+  void enablePopGesture({required String containerId}) {
     assert(containerId != null && containerId.isNotEmpty);
     BoostChannel.instance.sendEventToNative(containerId, {
       'event': 'enablePopGesture',
@@ -55,7 +54,7 @@ class BoostChannel {
   }
 
   /// disable iOS native pop gesture for container matching [containerId]
-  void disablePopGesture({@required String containerId}) {
+  void disablePopGesture({required String containerId}) {
     assert(containerId != null && containerId.isNotEmpty);
     BoostChannel.instance.sendEventToNative(containerId, {
       'event': 'enablePopGesture',
