@@ -6,13 +6,13 @@ import 'boost_container.dart';
 import 'container_overlay.dart';
 import 'flutter_boost_app.dart';
 
-typedef FlutterBoostRouteFactory = Route<dynamic> Function(
-    RouteSettings settings, String uniqueId);
+typedef FlutterBoostRouteFactory = Route<dynamic>? Function(
+    RouteSettings settings, String? uniqueId);
 
 FlutterBoostRouteFactory routeFactoryWrapper(
-    FlutterBoostRouteFactory routeFactory) {
+    FlutterBoostRouteFactory? routeFactory) {
   return (settings, uniqueId) {
-    var route = routeFactory(settings, uniqueId);
+    var route = routeFactory!(settings, uniqueId);
     if (route == null && settings.name == '/') {
       route = PageRouteBuilder<dynamic>(
           settings: settings, pageBuilder: (_, __, ___) => Container());
@@ -30,15 +30,15 @@ class BoostNavigator {
   static final BoostNavigator _instance = BoostNavigator._();
 
   /// The boost data center
-  FlutterBoostAppState appState;
+  FlutterBoostAppState? appState;
 
   /// The route table in flutter_boost
-  FlutterBoostRouteFactory _routeFactory;
+  FlutterBoostRouteFactory? _routeFactory;
 
-  set routeFactory(FlutterBoostRouteFactory routeFactory) =>
+  set routeFactory(FlutterBoostRouteFactory? routeFactory) =>
       _routeFactory = routeFactoryWrapper(routeFactory);
 
-  FlutterBoostRouteFactory get routeFactory => _routeFactory;
+  FlutterBoostRouteFactory? get routeFactory => _routeFactory;
 
   @Deprecated('Use `instance` instead.')
 
@@ -56,7 +56,7 @@ class BoostNavigator {
   /// If the name of route can be found in route table then return true,
   /// otherwise return false.
   bool isFlutterPage(String name) =>
-      routeFactory(RouteSettings(name: name), null) != null;
+      routeFactory!(RouteSettings(name: name), null) != null;
 
   /// Push the page with the given [name] onto the hybrid stack.
   /// [arguments] is the param you want to pass in next page
@@ -65,19 +65,19 @@ class BoostNavigator {
   /// if [opaque] is true,the page is opaque (not transparent)
   ///
   /// And it will return the result popped by page as a Future<T>
-  Future<T> push<T extends Object>(String name,
-      {Map<String, dynamic> arguments,
+  Future<T> push<T extends Object?>(String name,
+      {Map<String, dynamic>? arguments,
       bool withContainer = false,
       bool opaque = true}) {
     bool is_flutter_page = isFlutterPage(name);
     if (is_flutter_page && withContainer) {
       // 1. open flutter page with container
       // Intercepted in BoostFlutterRouterApi.pushRoute
-      return appState.pushWithResult(name,
-          arguments: arguments, withContainer: withContainer, opaque: opaque);
+      return appState!.pushWithResult(name,
+          arguments: arguments  ?? <String, dynamic>{}, withContainer: withContainer, opaque: opaque);
     } else {
       // 2. open native page or flutter page without container
-      return appState.pushWithInterceptor(
+      return appState!.pushWithInterceptor(
           name, false /* isFromHost */, is_flutter_page,
           arguments: arguments, withContainer: withContainer, opaque: opaque);
     }
@@ -87,8 +87,8 @@ class BoostNavigator {
   /// 1.Push a new page onto pageStack
   /// 2.remove(pop) previous page
   Future<T> pushReplacement<T extends Object>(String name,
-      {Map<String, dynamic> arguments, bool withContainer = false}) async {
-    final id = getTopPageInfo().uniqueId;
+      {Map<String, dynamic>? arguments, bool withContainer = false}) async {
+    final id = getTopPageInfo()!.uniqueId;
 
     final result =
         push(name, arguments: arguments, withContainer: withContainer);
@@ -96,36 +96,36 @@ class BoostNavigator {
     Future.delayed(const Duration(milliseconds: 100), () {
       remove(id);
     });
-    return result;
+    return result as FutureOr<T>;
   }
 
   /// Pop the top-most page off the hybrid stack.
-  Future<bool> pop<T extends Object>([T result]) async =>
-      await appState.popWithResult(result);
+  Future<bool> pop<T extends Object>([T? result]) async =>
+      await appState!.popWithResult(result);
 
   /// PopUntil page off the hybrid stack.
-  Future<void> popUntil({String route, String uniqueId}) async =>
-      appState.popUntil(route: route, uniqueId: uniqueId);
+  Future<void> popUntil({String? route, String? uniqueId}) async =>
+      appState!.popUntil(route: route, uniqueId: uniqueId);
 
   /// Remove the page with the given [uniqueId] from hybrid stack.
   ///
   /// This API is for backwards compatibility.
   /// Please use [BoostNavigator.pop] instead.
-  Future<bool> remove(String uniqueId,
-          {Map<String, dynamic> arguments}) async =>
-      appState.removeWithResult(uniqueId, arguments);
+  Future<bool> remove(String? uniqueId,
+          {Map<String, dynamic>? arguments}) async =>
+      appState!.removeWithResult(uniqueId, arguments);
 
   /// Retrieves the infomation of the top-most flutter page
   /// on the hybrid stack, such as uniqueId, pagename, etc;
   ///
   /// This is a legacy API for backwards compatibility.
-  PageInfo getTopPageInfo() => appState.getTopPageInfo();
+  PageInfo? getTopPageInfo() => appState!.getTopPageInfo();
 
   @Deprecated('use getPageInfoByContext(BuildContext context) instead')
-  PageInfo getTopByContext(BuildContext context) =>
+  PageInfo? getTopByContext(BuildContext context) =>
       BoostContainer.of(context)?.pageInfo;
 
-  PageInfo getPageInfoByContext(BuildContext context) =>
+  PageInfo? getPageInfoByContext(BuildContext context) =>
       BoostContainer.of(context)?.pageInfo;
 
   bool isTopPage(BuildContext context) {
@@ -135,15 +135,15 @@ class BoostNavigator {
   /// Return the number of flutter pages
   ///
   /// This is a legacy API for backwards compatibility.
-  int pageSize() => appState.pageSize();
+  int pageSize() => appState!.pageSize();
 }
 
 /// The PageInfo use in FlutterBoost ,it is not a public api
 class PageInfo {
   PageInfo({this.pageName, this.uniqueId, this.arguments, this.withContainer});
 
-  bool withContainer;
-  String pageName;
-  String uniqueId;
-  Map<String, dynamic> arguments;
+  bool? withContainer;
+  String? pageName;
+  String? uniqueId;
+  Map<String, dynamic>? arguments;
 }
