@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -20,12 +21,12 @@ typedef FlutterBoostAppBuilder = Widget Function(Widget home);
 class FlutterBoostApp extends StatefulWidget {
   FlutterBoostApp(
     FlutterBoostRouteFactory routeFactory, {
-    Key key,
-    FlutterBoostAppBuilder appBuilder,
-    String initialRoute,
+    Key? key,
+    FlutterBoostAppBuilder? appBuilder,
+    String? initialRoute,
 
     ///interceptors is to intercept push operation now
-    List<BoostInterceptor> interceptors,
+    List<BoostInterceptor>? interceptors,
   })  : appBuilder = appBuilder ?? _defaultAppBuilder,
         interceptors = interceptors ?? <BoostInterceptor>[],
         initialRoute = initialRoute ?? '/',
@@ -52,8 +53,8 @@ class FlutterBoostApp extends StatefulWidget {
 class FlutterBoostAppState extends State<FlutterBoostApp> {
   static const String _appLifecycleChangedKey = "app_lifecycle_changed_key";
 
-  final Map<String, Completer<Object>> _pendingResult =
-      <String, Completer<Object>>{};
+  final Map<String, Completer<Object?>> _pendingResult =
+      <String, Completer<Object?>>{};
 
   List<BoostContainer> get containers => _containers;
   final List<BoostContainer> _containers = <BoostContainer>[];
@@ -61,14 +62,14 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   /// All interceptors from widget
   List<BoostInterceptor> get interceptors => widget.interceptors;
 
-  BoostContainer get topContainer =>
+  BoostContainer? get topContainer =>
       containers.isNotEmpty ? containers.last : null;
 
-  NativeRouterApi get nativeRouterApi => _nativeRouterApi;
-  NativeRouterApi _nativeRouterApi;
+  NativeRouterApi? get nativeRouterApi => _nativeRouterApi;
+  NativeRouterApi? _nativeRouterApi;
 
-  BoostFlutterRouterApi get boostFlutterRouterApi => _boostFlutterRouterApi;
-  BoostFlutterRouterApi _boostFlutterRouterApi;
+  BoostFlutterRouterApi? get boostFlutterRouterApi => _boostFlutterRouterApi;
+  BoostFlutterRouterApi? _boostFlutterRouterApi;
 
   final Set<int> _activePointers = <int>{};
 
@@ -76,7 +77,7 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   final Map<String, List<EventListener>> _listenersTable =
       <String, List<EventListener>>{};
 
-  VoidCallback _lifecycleStateListenerRemover;
+  late VoidCallback _lifecycleStateListenerRemover;
 
   @override
   void initState() {
@@ -95,7 +96,7 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
 
     // Make sure that the widget in the tree that matches [overlayKey]
     // is already mounted, or [refreshOnPush] will fail.
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       // try to restore routes from host when hot restart.
       assert(() {
         _restoreStackForHotRestart();
@@ -103,7 +104,7 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
       }());
 
       refreshOnPush(initialContainer);
-      _boostFlutterRouterApi.isEnvReady = true;
+      _boostFlutterRouterApi!.isEnvReady = true;
       _addAppLifecycleStateEventListener();
       BoostOperationQueue.instance.runPendingOperations();
     });
@@ -123,10 +124,10 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
       final int index = arguments["lifecycleState"];
 
       if (index == AppLifecycleState.resumed.index) {
-        BoostFlutterBinding.instance
+        BoostFlutterBinding.instance!
             .changeAppLifecycleState(AppLifecycleState.resumed);
       } else if (index == AppLifecycleState.paused.index) {
-        BoostFlutterBinding.instance
+        BoostFlutterBinding.instance!
             .changeAppLifecycleState(AppLifecycleState.paused);
       }
       return;
@@ -143,9 +144,9 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   Widget build(BuildContext context) {
     return widget.appBuilder(WillPopScope(
         onWillPop: () async {
-          final canPop = topContainer.navigator.canPop();
+          final canPop = topContainer!.navigator!.canPop();
           if (canPop) {
-            topContainer.navigator.pop();
+            topContainer!.navigator!.pop();
             return true;
           }
           return false;
@@ -169,50 +170,50 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   }
 
   void _cancelActivePointers() {
-    _activePointers.toList().forEach(WidgetsBinding.instance.cancelPointer);
+    _activePointers.toList().forEach(WidgetsBinding.instance!.cancelPointer);
   }
 
-  String _createUniqueId(String pageName) {
+  String _createUniqueId(String? pageName) {
     return '${DateTime.now().millisecondsSinceEpoch}_$pageName';
   }
 
   BoostContainer _createContainer(PageInfo pageInfo) {
     pageInfo.uniqueId ??= _createUniqueId(pageInfo.pageName);
     return BoostContainer(
-        key: ValueKey<String>(pageInfo.uniqueId), pageInfo: pageInfo);
+        key: ValueKey<String?>(pageInfo.uniqueId), pageInfo: pageInfo);
   }
 
   Future<void> _saveStackForHotRestart() async {
     final stack = StackInfo();
-    stack.containers = <String>[];
+    stack.containers = <String?>[];
     for (var container in containers) {
-      stack.containers.add(container.pageInfo.uniqueId);
-      stack.routes = <String, List<Map<String, Object>>>{};
-      final params = <Map<String, Object>>[];
+      stack.containers!.add(container.pageInfo.uniqueId);
+      stack.routes = <String?, List<Map<String, Object?>>>{};
+      final List<Map<String, Object?>> params = [];
       for (var page in container.pages) {
-        final param = <String, Object>{};
+        final param = <String, Object?>{};
         param['pageName'] = page.pageInfo.pageName;
         param['uniqueId'] = page.pageInfo.uniqueId;
-        param['arguments'] = page.pageInfo.arguments;
+        param['arguments'] = page.pageInfo.arguments ?? {};
         params.add(param);
       }
-      stack.routes[container.pageInfo.uniqueId] = params;
+      stack.routes![container.pageInfo.uniqueId] = params;
     }
-    await nativeRouterApi.saveStackToHost(stack);
+    await nativeRouterApi!.saveStackToHost(stack);
     Logger.log(
-        '_saveStackForHotRestart, ${stack?.containers}, ${stack?.routes}');
+        '_saveStackForHotRestart, ${stack.containers}, ${stack.routes}');
   }
 
   Future<void> _restoreStackForHotRestart() async {
-    final stack = await nativeRouterApi.getStackFromHost();
+    final stack = await nativeRouterApi!.getStackFromHost();
     if (stack != null && stack.containers != null) {
-      for (String uniqueId in stack.containers) {
+      for (String uniqueId in stack.containers!.cast<String>()) {
         var withContainer = true;
-        final List<Object> routeList = stack.routes[uniqueId];
-        if (routeList != null) {
-          for (Map<Object, Object> route in routeList) {
-            var pageName = route['pageName'] as String;
-            var uniqueId = route['uniqueId'] as String;
+        final routeList = stack.routes![uniqueId];
+        if (routeList != null && routeList is List) {
+          for (var route in routeList) {
+            var pageName = route['pageName'] as String?;
+            var uniqueId = route['uniqueId'] as String?;
             var arguments = Map<String, dynamic>.from(
                 route['arguments'] ?? <String, dynamic>{});
             withContainer
@@ -225,25 +226,25 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
       }
     }
     Logger.log(
-        '_restoreStackForHotRestart, ${stack?.containers}, ${stack?.routes}');
+        '_restoreStackForHotRestart, ${stack.containers}, ${stack.routes}');
   }
 
-  Future<T> pushWithInterceptor<T extends Object>(
-      String name, bool isFromHost, bool isFlutterPage,
-      {Map<String, dynamic> arguments,
-      String uniqueId,
-      bool withContainer,
+  Future<T> pushWithInterceptor<T extends Object?>(
+      String? name, bool isFromHost, bool isFlutterPage,
+      {Map<String, dynamic>? arguments,
+      String? uniqueId,
+      required bool withContainer,
       bool opaque = true}) {
     Logger.log('pushWithInterceptor, uniqueId=$uniqueId, name=$name');
     var pushOption = BoostInterceptorOption(name,
         uniqueId: uniqueId,
         isFromHost: isFromHost,
         arguments: arguments ?? <String, dynamic>{});
-    var state = InterceptorState<BoostInterceptorOption>(pushOption);
+    InterceptorState state = InterceptorState<BoostInterceptorOption>(pushOption);
     for (var interceptor in interceptors) {
       final pushHandler = PushInterceptorHandler();
       interceptor.onPrePush(state.data, pushHandler);
-      state = pushHandler.state;
+      state = pushHandler.state!;
       if (state.type != InterceptorResultType.next) {
         Logger.log('The page was intercepted by user. name:$name, '
             'isFromHost=$isFromHost, isFlutterPage=$isFlutterPage');
@@ -270,7 +271,7 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
           final params = CommonParams()
             ..pageName = pushOption.name
             ..arguments = pushOption.arguments;
-          nativeRouterApi.pushNativeRoute(params);
+          nativeRouterApi!.pushNativeRoute(params);
           return pendNativeResult(pushOption.name);
         }
       }
@@ -278,13 +279,13 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
       Logger.log(
           'Oops, Unrecognized parameter type: ${state.data.runtimeType}');
     }
-    return Future<void>.value();
+    return Future<T>.value();
   }
 
-  Future<T> pushWithResult<T extends Object>(String pageName,
-      {String uniqueId,
-      Map<String, dynamic> arguments,
-      bool withContainer,
+  Future<T> pushWithResult<T extends Object?>(String? pageName,
+      {String? uniqueId,
+      Map<String, dynamic>? arguments,
+      required bool withContainer,
       bool opaque = true}) {
     uniqueId ??= _createUniqueId(pageName);
     if (withContainer) {
@@ -293,8 +294,8 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
         ..pageName = pageName
         ..uniqueId = uniqueId
         ..opaque = opaque
-        ..arguments = arguments ?? <String, dynamic>{};
-      nativeRouterApi.pushFlutterRoute(params);
+        ..arguments = arguments;
+      nativeRouterApi!.pushFlutterRoute(params);
       _pendingResult[uniqueId] = completer;
       return completer.future;
     } else {
@@ -302,8 +303,8 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
     }
   }
 
-  Future<T> pushPage<T extends Object>(String pageName,
-      {String uniqueId, Map<String, dynamic> arguments}) {
+  Future<T> pushPage<T extends Object?>(String? pageName,
+      {String? uniqueId, Map<String, dynamic>? arguments}) {
     Logger.log('pushPage, uniqueId=$uniqueId, name=$pageName,'
         ' arguments:$arguments, $topContainer');
     final pageInfo = PageInfo(
@@ -312,19 +313,19 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
         arguments: arguments,
         withContainer: false);
     assert(topContainer != null);
-    var result = topContainer.addPage(BoostPage.create(pageInfo));
     _pushFinish(pageName, uniqueId: uniqueId, arguments: arguments);
-    return result;
+    return topContainer!.addPage(BoostPage.create(pageInfo));
   }
 
-  void pushContainer(String pageName,
-      {String uniqueId,
+
+  void pushContainer(String? pageName,
+      {String? uniqueId,
       bool isFromHost = false,
-      Map<String, dynamic> arguments}) {
+      Map<String, dynamic>? arguments}) {
     _cancelActivePointers();
     final existed = _findContainerByUniqueId(uniqueId);
     if (existed != null) {
-      if (topContainer?.pageInfo?.uniqueId != uniqueId) {
+      if (topContainer?.pageInfo.uniqueId != uniqueId) {
         containers.remove(existed);
         containers.add(existed);
 
@@ -353,40 +354,40 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
         ' arguments:$arguments, $containers');
   }
 
-  void _pushFinish(String pageName,
-      {String uniqueId,
-      bool isFromHost = false,
-      Map<String, dynamic> arguments}) {
+  void _pushFinish(String? pageName,
+      {String? uniqueId,
+        bool isFromHost = false,
+        Map<String, dynamic>? arguments}) {
     var pushOption = BoostInterceptorOption(pageName,
         uniqueId: uniqueId,
         isFromHost: isFromHost,
         arguments: arguments ?? <String, dynamic>{});
-    var state = InterceptorState<BoostInterceptorOption>(pushOption);
+    InterceptorState state = InterceptorState<BoostInterceptorOption>(pushOption);
     for (var interceptor in interceptors) {
       final pushHandler = PushInterceptorHandler();
       interceptor.onPostPush(state.data, pushHandler);
-      state = pushHandler.state;
+      state = pushHandler.state!;
       if (state.type != InterceptorResultType.next) {
         break;
       }
     }
   }
 
-  Future<bool> popWithResult<T extends Object>([T result]) async {
-    final uniqueId = topContainer?.topPage?.pageInfo?.uniqueId;
+  Future<bool> popWithResult<T extends Object>([T? result]) async {
+    final uniqueId = topContainer?.topPage.pageInfo.uniqueId;
     _completePendingResultIfNeeded(uniqueId, result: result);
     return await pop(result: result);
   }
 
   Future<bool> removeWithResult(
-      [String uniqueId, Map<String, dynamic> result]) async {
+      [String? uniqueId, Map<String, dynamic>? result]) async {
     _completePendingResultIfNeeded(uniqueId, result: result);
     return await pop(uniqueId: uniqueId, result: result);
   }
 
-  void popUntil({String route, String uniqueId}) async {
-    BoostContainer targetContainer;
-    BoostPage targetPage;
+  void popUntil({String? route, String? uniqueId}) async {
+    BoostContainer? targetContainer;
+    BoostPage? targetPage;
     int popUntilIndex = containers.length;
     if (uniqueId != null) {
       for (int index = containers.length - 1; index >= 0; index--) {
@@ -434,23 +435,23 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
           ..pageName = container.pageInfo.pageName
           ..uniqueId = container.pageInfo.uniqueId
           ..arguments = {"animated": false};
-        await nativeRouterApi.popRoute(params);
+        await nativeRouterApi!.popRoute(params);
       }
 
       if (targetContainer.topPage != targetPage) {
         Future<void>.delayed(
             const Duration(milliseconds: 50),
             () => targetContainer?.navigator
-                ?.popUntil(ModalRoute.withName(targetPage.name)));
+                ?.popUntil(ModalRoute.withName(targetPage!.name!)));
       }
     } else {
-      topContainer?.navigator?.popUntil(ModalRoute.withName(targetPage.name));
+      topContainer?.navigator?.popUntil(ModalRoute.withName(targetPage!.name!));
     }
   }
 
   Future<bool> pop(
-      {String uniqueId, Object result, bool onBackPressed = false}) async {
-    BoostContainer container;
+      {String? uniqueId, Object? result, bool onBackPressed = false}) async {
+    BoostContainer? container;
     if (uniqueId != null) {
       container = _findContainerByUniqueId(uniqueId);
       if (container == null) {
@@ -465,7 +466,7 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
       container = topContainer;
     }
 
-    final currentPage = topContainer?.topPage?.pageInfo?.uniqueId;
+    final String currentPage = topContainer?.topPage.pageInfo.uniqueId ?? "";
     assert(currentPage != null);
     _completePendingResultIfNeeded(currentPage);
 
@@ -475,29 +476,28 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
     // 3.If uniqueId is not topPage's uniqueId, so we will remove an existing
     // page in container.
     if (uniqueId == null ||
-        uniqueId == container.pages.last.pageInfo.uniqueId) {
+        uniqueId == container!.pages.last.pageInfo.uniqueId) {
       final handled = onBackPressed
-          ? await _performBackPressed(container, result)
+          ? await _performBackPressed(container!, result)
           : container?.navigator?.canPop();
       if (handled != null) {
         if (!handled) {
-          assert(container.pageInfo.withContainer);
+          assert(container!.pageInfo.withContainer!);
           final params = CommonParams()
-            ..pageName = container.pageInfo.pageName
+            ..pageName = container!.pageInfo.pageName
             ..uniqueId = container.pageInfo.uniqueId
             ..arguments =
-                (result is Map<String, dynamic>) ? result : <String, dynamic>{};
-          await nativeRouterApi.popRoute(params);
+            (result is Map<String, dynamic>) ? result : <String, dynamic>{};
+          await nativeRouterApi!.popRoute(params);
         } else {
           if (!onBackPressed) {
-            container.navigator.pop(result);
+            container!.navigator!.pop(result);
           }
         }
       }
     } else {
-      final page = container.pages.singleWhere(
-          (entry) => entry.pageInfo.uniqueId == uniqueId,
-          orElse: () => null);
+      final page = container.pages.singleWhereOrNull(
+          (entry) => entry.pageInfo.uniqueId == uniqueId);
       container.removePage(page);
     }
 
@@ -506,39 +506,39 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   }
 
   Future<bool> _performBackPressed(
-      BoostContainer container, Object result) async {
-    if (container?.backPressedHandler != null) {
-      container.backPressedHandler.call();
+      BoostContainer container, Object? result) async {
+    if (container.backPressedHandler != null) {
+      container.backPressedHandler!.call();
       return true;
     } else {
-      return await container?.navigator?.maybePop(result);
+      return (await container.navigator?.maybePop(result))!;
     }
   }
 
   Future<void> _removeContainer(BoostContainer container) async {
-    if (container.pageInfo.withContainer) {
+    if (container.pageInfo.withContainer!) {
       Logger.log('_removeContainer ,  uniqueId=${container.pageInfo.uniqueId}');
       final params = CommonParams()
         ..pageName = container.pageInfo.pageName
         ..uniqueId = container.pageInfo.uniqueId
-        ..arguments = container.pageInfo.arguments;
-      return await _nativeRouterApi.popRoute(params);
+        ..arguments = Map<String, dynamic>.from(container.pageInfo.arguments ?? <String, dynamic>{});
+      return await _nativeRouterApi!.popRoute(params);
     }
   }
 
   void onForeground() {
     if (topContainer != null) {
-      BoostLifecycleBinding.instance.appDidEnterForeground(topContainer);
+      BoostLifecycleBinding.instance.appDidEnterForeground(topContainer!);
     }
   }
 
   void onBackground() {
     if (topContainer != null) {
-      BoostLifecycleBinding.instance.appDidEnterBackground(topContainer);
+      BoostLifecycleBinding.instance.appDidEnterBackground(topContainer!);
     }
   }
 
-  BoostContainer _findContainerByUniqueId(String uniqueId) {
+  BoostContainer? _findContainerByUniqueId(String? uniqueId) {
     //Because first page can be removed from container.
     //So we find id in container's PageInfo
     //If we can't find a container matching this id,
@@ -547,21 +547,19 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
     //
     //If we can't find any container or page matching this id,we return null
 
-    var result = containers.singleWhere(
-        (element) => element.pageInfo.uniqueId == uniqueId,
-        orElse: () => null);
+    var result = containers.singleWhereOrNull(
+        (element) => element.pageInfo.uniqueId == uniqueId);
 
     if (result != null) {
       return result;
     }
 
-    return containers.singleWhere(
+    return containers.singleWhereOrNull(
         (element) => element.pages
-            .any((element) => element.pageInfo.uniqueId == uniqueId),
-        orElse: () => null);
+            .any((element) => element.pageInfo.uniqueId == uniqueId));
   }
 
-  void remove(String uniqueId) {
+  void remove(String? uniqueId) {
     if (uniqueId == null) {
       return;
     }
@@ -575,18 +573,17 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
       refreshOnRemove(container);
     } else {
       for (var container in containers) {
-        final page = container.pages.singleWhere(
-            (entry) => entry.pageInfo.uniqueId == uniqueId,
-            orElse: () => null);
+        final page = container.pages.singleWhereOrNull(
+            (entry) => entry.pageInfo.uniqueId == uniqueId);
         container.removePage(page);
       }
     }
     Logger.log('remove,  uniqueId=$uniqueId, $containers');
   }
 
-  Future<T> pendNativeResult<T extends Object>(String pageName) {
+  Future<T> pendNativeResult<T extends Object?>(String? pageName) {
     final completer = Completer<T>();
-    final initiatorPage = topContainer?.topPage?.pageInfo?.uniqueId;
+    final initiatorPage = topContainer?.topPage.pageInfo.uniqueId;
     final key = '$initiatorPage#$pageName';
     _pendingResult[key] = completer;
     Logger.log('pendNativeResult, key:$key, size:${_pendingResult.length}');
@@ -594,26 +591,26 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   }
 
   void onNativeResult(CommonParams params) {
-    final initiatorPage = topContainer?.topPage?.pageInfo?.uniqueId;
+    final initiatorPage = topContainer?.topPage.pageInfo.uniqueId;
     final key = '$initiatorPage#${params.pageName}';
     if (_pendingResult.containsKey(key)) {
-      _pendingResult[key].complete(params.arguments);
+      _pendingResult[key]!.complete(params.arguments ?? {});
       _pendingResult.remove(key);
     }
     Logger.log('onNativeResult, key:$key, result:${params.arguments}');
   }
 
-  void _completePendingResultIfNeeded<T extends Object>(String uniqueId,
-      {T result}) {
+  void _completePendingResultIfNeeded<T extends Object>(String? uniqueId,
+      {T? result}) {
     if (uniqueId != null && _pendingResult.containsKey(uniqueId)) {
-      _pendingResult[uniqueId].complete(result);
+      _pendingResult[uniqueId]!.complete(result ?? {});
       _pendingResult.remove(uniqueId);
     }
   }
 
   void onContainerShow(CommonParams params) {
     final container = _findContainerByUniqueId(params.uniqueId);
-    BoostLifecycleBinding.instance.containerDidShow(container);
+    BoostLifecycleBinding.instance.containerDidShow(container!);
   }
 
   void onContainerHide(CommonParams params) {
@@ -628,8 +625,8 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
   ///Calls when Native send event to flutter(here)
   void onReceiveEventFromNative(CommonParams params) {
     //Get the name and args from native
-    var key = params.key;
-    Map args = params.arguments;
+    var key = params.key!;
+    Map? args = params.arguments;
     assert(key != null);
 
     //Get all of listeners matching this key
@@ -638,7 +635,7 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
     if (listeners == null) return;
 
     for (final listener in listeners) {
-      listener(key, args);
+      listener(key, args ?? {});
     }
   }
 
@@ -655,13 +652,13 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
     listeners.add(listener);
 
     return () {
-      listeners.remove(listener);
+      listeners!.remove(listener);
     };
   }
 
   ///Interal methods below
 
-  PageInfo getTopPageInfo() {
+  PageInfo? getTopPageInfo() {
     return topContainer?.topPage?.pageInfo;
   }
 
@@ -707,19 +704,19 @@ class FlutterBoostAppState extends State<FlutterBoostApp> {
 
 // ignore: must_be_immutable
 class BoostPage<T> extends Page<T> {
-  BoostPage({LocalKey key, this.pageInfo})
+  BoostPage({LocalKey? key, required this.pageInfo})
       : super(key: key, name: pageInfo.pageName, arguments: pageInfo.arguments);
   final PageInfo pageInfo;
 
   static BoostPage<dynamic> create(PageInfo pageInfo) {
     final page = BoostPage<dynamic>(key: UniqueKey(), pageInfo: pageInfo);
-    page._route = BoostNavigator.instance.routeFactory(page, pageInfo.uniqueId);
+    page._route = BoostNavigator.instance.routeFactory!(page, pageInfo.uniqueId);
     return page;
   }
 
-  Route<T> _route;
+  Route<T>? _route;
 
-  Route<T> get route => _route;
+  Route<T>? get route => _route;
 
   /// A future that completes when this page is popped.
   Future<T> get popped => _popCompleter.future;
@@ -737,7 +734,7 @@ class BoostPage<T> extends Page<T> {
 
   @override
   Route<T> createRoute(BuildContext context) {
-    return _route;
+    return _route!;
   }
 }
 
@@ -745,10 +742,10 @@ class BoostNavigatorObserver extends NavigatorObserver {
   BoostNavigatorObserver();
 
   @override
-  void didPush(Route<dynamic> route, Route<dynamic> previousRoute) {
+  void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
     //handle internal route but ignore dialog or abnormal route.
     //otherwise, the normal page will be affected.
-    if (previousRoute != null && route?.settings?.name != null) {
+    if (previousRoute != null && route.settings.name != null) {
       BoostLifecycleBinding.instance.routeDidPush(route, previousRoute);
     }
 
@@ -763,8 +760,8 @@ class BoostNavigatorObserver extends NavigatorObserver {
   }
 
   @override
-  void didPop(Route<dynamic> route, Route<dynamic> previousRoute) {
-    if (previousRoute != null && route?.settings?.name != null) {
+  void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) {
+    if (previousRoute != null && route.settings.name != null) {
       BoostLifecycleBinding.instance.routeDidPop(route, previousRoute);
     }
 
@@ -780,7 +777,7 @@ class BoostNavigatorObserver extends NavigatorObserver {
   }
 
   @override
-  void didRemove(Route<dynamic> route, Route<dynamic> previousRoute) {
+  void didRemove(Route<dynamic> route, Route<dynamic>? previousRoute) {
     final navigatorObserverList =
         BoostLifecycleBinding.instance.navigatorObserverList;
     if (navigatorObserverList != null && navigatorObserverList.isNotEmpty) {
@@ -795,7 +792,7 @@ class BoostNavigatorObserver extends NavigatorObserver {
   }
 
   @override
-  void didReplace({Route<dynamic> newRoute, Route<dynamic> oldRoute}) {
+  void didReplace({Route<dynamic>? newRoute, Route<dynamic>? oldRoute}) {
     final navigatorObserverList =
         BoostLifecycleBinding.instance.navigatorObserverList;
     if (navigatorObserverList != null && navigatorObserverList.isNotEmpty) {
@@ -807,7 +804,7 @@ class BoostNavigatorObserver extends NavigatorObserver {
   }
 
   @override
-  void didStartUserGesture(Route<dynamic> route, Route<dynamic> previousRoute) {
+  void didStartUserGesture(Route<dynamic> route, Route<dynamic>? previousRoute) {
     final navigatorObserverList =
         BoostLifecycleBinding.instance.navigatorObserverList;
     if (navigatorObserverList != null && navigatorObserverList.isNotEmpty) {
