@@ -1,16 +1,16 @@
 import 'package:pigeon/pigeon.dart';
 
 class CommonParams {
-  String pageName;
-  String uniqueId;
-  Map<String, Object> arguments;
-  bool opaque;
-  String key;
+  String? pageName;
+  String? uniqueId;
+  Map<String, Object>? arguments;
+  bool? opaque;
+  String? key;
 }
 
 class StackInfo {
-  List<String> containers;
-  Map<String, List<Map<String, Object>>> routes;
+  List<String>? containers;
+  Map<String, List<Map<String, Object>>>? routes;
 }
 
 @HostApi()
@@ -42,7 +42,7 @@ void configurePigeon(PigeonOptions opts) {
   opts.dartOut = 'lib/src/messages.dart';
   opts.objcHeaderOut = 'ios/Classes/messages.h';
   opts.objcSourceOut = 'ios/Classes/messages.m';
-  opts.objcOptions.prefix = 'FB';
+  opts.objcOptions!.prefix = 'FB';
   opts.javaOut =
       'android/src/main/java/com/idlefish/flutterboost/Messages.java';
 }
